ATM: add time field User: add getPassword() Transaction: change type of `amount` from float to double Account: specify more clearly about the type of Exception threw InvalidOperationException: add one type of Exception

+++ phase1/src/atm/Transaction.java
@@ -3,9 +3,9 @@ package atm;
 public class Transaction {
 	private Account source;
 	private Account destination;
-	private float amount;
+	private double amount;
 	
-	public Transaction(Account source, Account destination, float amount) {
+	public Transaction(Account source, Account destination, double amount) {
 		this.source = source;
 		this.destination = destination;
 		this.amount = amount;
